Refactor generic message provider logic into the MessageProviderBase class. Refactor configuration class to the same form as in sensenet.
