+++ src/SenseNet.Security.Tests/MessagingTests.cs
@@ -58,7 +58,7 @@ namespace SenseNet.Security.Tests
             var smallActivityId = smallActivity.Id;
 
             // large activity
-            var largeActivity = new TestActivity() { Body = new String('*', Configuration.DistributableSecurityActivityMaxSize + 1) };
+            var largeActivity = new TestActivity() { Body = new String('*', Configuration.Messaging.DistributableSecurityActivityMaxSize + 1) };
             largeActivity.Execute(context.Security);
             var largeActivityId = largeActivity.Id;
