+++ src/SenseNet.Security/Configuration/Identities.cs
@@ -1,0 +1,10 @@
+﻿namespace SenseNet.Security.Configuration
+{
+    internal static class Identities
+    {
+        public static int SystemUserId { get; internal set; }
+        public static int VisitorUserId { get; internal set; }
+        public static int EveryoneGroupId { get; internal set; }
+        public static int OwnerGroupId { get; internal set; }
+    }
+}
