+++ src/SenseNet.Security/SenseNet.Security.csproj
@@ -64,7 +64,8 @@
     <Compile Include="AceInfo.cs" />
     <Compile Include="AclEditor.cs" />
     <Compile Include="AclInfo.cs" />
-    <Compile Include="Configuration.cs" />
+    <Compile Include="Configuration\Messaging.cs" />
+    <Compile Include="Configuration\Identities.cs" />
     <Compile Include="DataHandler.cs" />
     <Compile Include="Data\DatabaseStorage.cs" />
     <Compile Include="Data\MemoryDataProvider.cs" />
@@ -75,6 +76,7 @@
     <Compile Include="ISecurityDataProvider.cs" />
     <Compile Include="ISecurityUser.cs" />
     <Compile Include="Messaging\CommunicationMonitor.cs" />
+    <Compile Include="Messaging\MessageProviderBase.cs" />
     <Compile Include="Messaging\SecurityActivityHistory.cs" />
     <Compile Include="Messaging\SecurityActivityLoader.cs" />
     <Compile Include="Messaging\SecurityMessages\AddUserToSecurityGroupsActivity.cs" />
